Some further improvements to AlefPacketInterface, mainly in that we now handle strings instead of char arrays LoginCharPos is now handled within the data system authToken support has been added, SQL will be updated soon to reflect this change. Removed worldAddress from AlefLogin.ini, as we now request it from the serverListSys Did some cleaning of AlefLogin AlefWorld has been cleaned and prepared to be worked on, including taking in a given authToken and knowing which account it is tied to, the next step is to then pull character data from that account using the known account id and the character name provided by the client Other minor various cleanings/cosmetic fixes.

+++ AlefCommon/AlefClientCharacter.h
@@ -14,7 +14,7 @@ public:
 	void setCurChar(CharacterData charData) { curChar = charData; }
 	CharacterData getCurrentChar() { return curChar; }
 	CharacterData getCharAtSlot(Int32 slotIdx);
-	UInt32 getCharCount() { return characters.size(); }
+	UInt32 getCharCount() { return (UInt32)characters.size(); }
 	UInt32 getAvailableSlotNum();
 
 	void setAcctID(UInt32 ID) { acctID = ID; }
